Added scaling feature to viewer.
git-svn-id: http://kos.informatik.uni-osnabrueck.de/software/meshing@2369 bea721bf-f9f1-0310-a2d5-e0bfdc8dd81e

+++ src/viewer/ObjectHandler.h
@@ -35,7 +35,7 @@ public:
 	void objectSelected();
 
 	void transformSelectedObject(Matrix4 m);
-	void transformSelectedObject(int mode, int d);
+	void transformSelectedObject(int mode, double d);
 
 	void printTransformation(){cout << objects[selectedObject]->getTransformation();};
